extract out remaining functions without too much refactoring
