+++ src/functions/inputProntoReceiptIntoMagento.ts
@@ -1,0 +1,133 @@
+import { Page } from "puppeteer";
+import { orderWithMagCommentResult, orderWithSellResult } from "../types.js";
+import { saveContent, waitTillHTMLRendered } from "./utils.js";
+
+export default async function inputProntoReceiptIntoMagento(
+  magentoPage: Page,
+  order: orderWithSellResult,
+): Promise<orderWithMagCommentResult> {
+  if (!Object.prototype.hasOwnProperty.call(order, "result")) {
+    throw new Error(
+      "wrong argument passed as parameter to inputProntoReceiptIntoMagento function",
+    );
+  }
+
+  console.log("input recept in mag func running on ", order);
+  // nav to order search page
+  await magentoPage.goto(
+    `https://www.paddypallin.com.au/agpallin_20/sales/order/index/key/${process.env.MAG_KEY}/`,
+  );
+  await magentoPage.waitForSelector(
+    "input.admin__control-text.data-grid-search-control",
+  );
+  // try to remove active filters
+  console.log("try to remove active filters");
+  try {
+    await magentoPage.click("button.action-remove");
+  } catch {
+    console.log("no filters to remove");
+  }
+  // clear input
+  const searchInput = await magentoPage.$(
+    "input.admin__control-text.data-grid-search-control",
+  );
+  await searchInput?.evaluate((input) => (input.value = ""));
+
+  // search for order and navigate to order detail page
+  await magentoPage.type(
+    "input.admin__control-text.data-grid-search-control",
+    order.magentoOrder,
+  );
+
+  await magentoPage.keyboard.press("Enter");
+  // how to wait for spinner to be display none
+  await magentoPage.waitForSelector("div.admin__data-grid-loading-mask", {
+    hidden: true,
+  });
+
+  await waitTillHTMLRendered(magentoPage);
+
+  await Promise.all([
+    magentoPage.waitForSelector("tr.data-row"),
+    magentoPage.waitForSelector("a.action-menu-item"),
+  ]);
+
+  await waitTillHTMLRendered(magentoPage);
+
+  // i need a better wait.as another idea i could extract out this function and try again until we get to the right page
+  // /*
+  // wait for the first element in the search results table to equal the search input
+  try {
+    await magentoPage.waitForSelector(
+      `::-p-xpath(/html/body/div[2]/main/div[2]/div/div/div/div[4]/table/tbody/tr[1]/td[2]/div[contains(text(),"${order.magentoOrder}")])`,
+    );
+  } catch (err) {
+    console.log("did not find the element via xpath");
+    console.error(err);
+    // throw new Error("did not wait for page navigation");
+  }
+  // */
+  const screenBeforeClick = await magentoPage.content();
+  await saveContent(magentoPage, screenBeforeClick, "screenBeforeClick");
+
+  // should find a better way of clicking?
+  // click on the row that has the order number?
+  // a click on any element in the row will take user to the order page?
+  // how do i select the parent element based upon child innerText,
+  // then click on the a tag in that parent
+
+  await Promise.all([
+    magentoPage.click("td > a"),
+    magentoPage.waitForNavigation(),
+  ]);
+
+  const screenAfterClick = await magentoPage.content();
+  await saveContent(magentoPage, screenAfterClick, "screenAfterClick");
+
+  // Check if we have navigated to correct page
+
+  const magOrderNumberFromPage = await magentoPage.$eval(
+    "h1.page-title",
+    (el) => {
+      return el.innerText.replace("#", "");
+    },
+  );
+  if (magOrderNumberFromPage !== order.magentoOrder) {
+    console.log(orderNumbers);
+    throw new Error(
+      "script is on the wrong page! error somewhere when navigating to the order detail page",
+    );
+  }
+
+  await magentoPage.waitForSelector("textarea#history_comment");
+  await magentoPage.type(
+    "textarea#history_comment",
+    `${order.prontoReceipt} - ${order.result}`,
+  );
+  // submit the comment
+  await magentoPage.click('button[title="Submit Comment"]');
+  // check if comment was sent successfully
+  await magentoPage.waitForSelector("div.loading-mask", { hidden: true });
+
+  await waitTillHTMLRendered(magentoPage);
+  await magentoPage.waitForSelector("ul.note-list");
+  const comments = await magentoPage.$$eval("div.note-list-comment", (el) => {
+    return el.map((comment) => {
+      console.log(comment.innerText);
+      return comment.innerText;
+    });
+  });
+
+  console.log(JSON.stringify(comments));
+
+  //what should this function return? how should we handle errors / make user aware of erros?
+  // e.g. network drops out during selling.
+
+  if (!JSON.stringify(comments).includes(order.prontoReceipt)) {
+    console.log("did not find order comment ");
+    return Promise.reject(
+      `did not find order comments for${order.magentoOrder} `,
+    );
+  }
+  return { ...order, magResult: "comment was made in magento" };
+}
