+++ src/functions/waitAndDisplayNeo.ts
@@ -1,0 +1,39 @@
+export default async function waitAndDisplayNeo(seconds: number) {
+  const milliseonds = seconds * 1000;
+
+  await new Promise((r) => setTimeout(r, milliseonds));
+  console.log(`xkkxxxkxxxkkxkkxkkxxxkkxxkl.   .';;;,,;:'..,:::okdodddkkkkkOXWWMMWKxdooxxxddxxxxxxxkkkOOkxdxkkddxkxx
+    xkxxxxxxxxxkxkxxkxxxkkxxxkl.....,::,,'';'..,:llxkddddxkOOO0KNWWMMMNxoddxxxddxxdxxxxkkkkOkxddxxddxxxx
+    xxxdxxxxdxxkkkkxkxxxxxxxxxd;...';;'.........,cldxooolloodxkOOKXNNWW0xKKkxxddxxdddxxkkxxkxxxdxxxxxxxd
+    dxdddxxdddxxkkkkkxxxxxxdxxko,.....          ..'cc;:;.........'ldx00KXWKxxxddxdddodxxkxxxxxxdxxxxxdxx
+  dxxdodxdddxxkkkkkxxxxxxdddxxl'.'.              ',,;.         .;,,okKNWKxddddxdddodxxxxxxxxxdxxxxxxxd
+  dxxdddddddxxxxxxxxddxxddddolo:;:.              '::c'         .;,'dXNKKKkddddxdddddddxxxxxxdddxkxxdoo
+  dxxdxxxddddxxdddxddddddddoccocco;.            .;dkOo'.       .clxXWWX0KOdoodddddddddxxxxxxdddxxdoood
+  dddodxxxdddxddddddddddddddc:ccldl;'..        ..ck0K0dc;....';o0NWMMWNNKxdooodddddddddddxxxddoooooddd
+  xdoooddddodddooddddddddddxxo::cdo:;,,,........,lk0K0Oxoc:loxk0XNWMWWWXxodoooodddddddddddxxddlloddddd
+  xxooddddooddoooddddddddddxddoc:odl:;;;,......',lk0K0OOkkkOkkkO0KNMWNXkoooolloddooddddxddxxxdlclddddd
+  ddoodxxdoodddoodddddddddddddddoodoc::;;,'....';lkKKKKKxdxkkkOOOKNWNKxooloollloooooddxxxdxxxdollodxxd
+  odoodxxdooddooodddoodoooddddxxddoolcc:;;,'.',,;cdO000XOloxkO0XXKNWKdoooololllloooodxxxddxkxdollodxxd
+  oolloddoooddooodooooooooodddddoooolccc:;,,'''.';::;,;dOxddkKNWWWWNOooooooollcloooodxxddodxxdlllodxxd
+  lllloooolooooloddooooooddddxxdoooollcc:;,''......',;okOOOxk0NWWWWKxooooollllcloolodxxdoodxxocclodxxd
+  clllooollloolloddooooooodddxxdoooolllc:;,'''...,';coxxxxxOOOXMWWXkooodolllllllollodxdolodxxoccclodoo
+  ccllolllllllllooooooloooddddxoloodollc:;,'''...,''',;::ccokOXWWWKdooooolllllcllolodxdoloddolcccclloo
+  ::clllccloolloooollloooooddoooooooollc:,'.'.........',:ccclxXWWWKdoooollllllcclllooddolloolcc:cclllo
+  ::lllccclooooooooolloooooddooloooolcclc;,,;;,'..'',,;cldxxdkXWWWXkolooccclolcclllloooolooolc:cccclll
+  :cllcclllloooooooollodooodooollol,';:cc:;,;;,''..',;:clodxOKXNWWWKdlolccclocclllllooolloool::ccccllo
+  llllllllloooooooooooodolooooollc'  .':;,;,;;,'.',:cldxk0KK0O0NNXXNKdoollllolccclllooolcoool::cccccoo
+  llclllllllloooooooooodoloolllol'     ...'',;,'',,;;:coxO0OkkO0OkOKN0dollcloolccllllllllllolc::cccclo
+  llcclcllllloooooooooddoooooool'         ...''.''''''',lO0kxdooodxOKNKxoooooollllolclllllllocccclccod
+  ,,,,;;;::::ccccclc:clllcloodo'            ....''....';dOxl;,;:cloxOKNKdolclolcccc:::::cc::c;;;:c;:oo
+  ,,',;;;;;;:::ccccc:cloool:;:;               ...''...',;,....';:cldx0XNX0kdolc::::;;,,,;:;;;,'',;,,cc
+  ''',;;,,;;;:clollclooc;..  ..                 ........     ..';:codk0K0XNNX0kdlc:;;,,;;;,,,,,,;;,,::
+  ,'',,,;cooddddooc,..                                        ..,:clok00O0KXXXNNK0koc;;;;;;;,,;;;;,,;:
+  :;;clllllc:::;;;'.                                          ..';:cok0OOO0KKKKXNNNNX0Okdl:;;;;;;,,,;:
+  lc:;'..........                                              ..';cdkOkOO000KKKKKXXXXNNNX0kdc:::;,,::
+  ..                                                            .,:ldkkkkO0OOOOO000000KKXNNWNX0kdc;;:;
+  ..',,:olloodxxxdoxxxxkkkOO0KXXNNWNKOdl:
+  .  ...'::,;cccllccllooddddxxxkO00KXNNNKk
+  .                        .. .;,',;:ccc:::::lllcclodddxxkkO0KNN
+  .                       ..  .,'.'',,;:;,',;cc;;;cloodddddxxk0O
+  ..           .           .   .'......',''.'':c'.,;:cllddooodxdl`);
+}
